Avoid compiler warnings about implicit function declarations.
Include header that declares functions from the UMFPACK library before
using them. Add path to that header to the preprocessor flags.
Also keep similar order of linker flags independent of whether the
bundled or an external UMFPACK library is used.

This warning is elevated to an error in LLVM Clang 18 (with default
compiler flags).