Fix METIS compilation error on MinGW (gettimeofday issue)
This commit resolves the compilation error in the METIS library when building
Elmer FEM on Windows using MinGW. The error was caused by the use of the
POSIX function \`gettimeofday\`, which is not directly available in Windows.

Changes made:
1. Modified \`elmergrid/src/metis-5.1.0/GKlib/gk_arch.h\` to define
   \`gettimeofday\` as \`mingw_gettimeofday\` when compiling with MinGW.
2. Updated \`elmergrid/src/metis-5.1.0/GKlib/timers.c\` to include \`gk_arch.h\`.

These changes allow the METIS library to compile successfully on Windows
with MinGW, resolving the \"implicit declaration of function 'gettimeofday'\"
error.

Fixes #489