Fix compatibility with newer versions of UMFPACK.
In old versions of UMFPACK (like the one that is bundled with ElmerFEM),
the `*l_*` versions of its functions used `long` for indices in their
interface.

That was later changed to use `SuiteSparse_long`. `SuiteSparse_long` was
`long long` on Windows and `long` on any other platform.

For the latest versions of UMFPACK, `SuiteSparse_long` is `int64_t` (on
all platforms).

This change fixes compilation errors on Windows (where `long` is 32-bit
wide also on 64-bit, LLP64) and macOS (where `int64_t` is `long long`
which has the same size as `long` but is still technically a different
type than `long`).
The change still maintains compatibility with old versions of UMFPACK
(like the one that is bundled with ElmerFEM).