Avoid out-of-bound index.
The test `rotflow` occasionally fails and valgrind flags an out-of-range
error when running that test.

That error occurs because bubble degrees of freedom with indices larger
than the number of nodes are used. Avoid that error by optionally not
including the indices of bubble degrees of freedom in `mGetElementDOFs`.
Only use non-bubble DOFs in `UpdateGlobalEquations`.