Check during configuration if `MPI_IN_PLACE` is working.
MSMPI is set up such that some variables that should be shared between
the MPI library and the program that is using it are dllimported via
variables in common blocks. While that is likely working with DEC or
Intel Fortran compilers on Windows, it doesn't work with GFortran.
See: https://gcc.gnu.org/bugzilla/show_bug.cgi?id=47030

Afaict, that doesn't only affect MSMPI but all MPI implementations on
Windows that are derived from MPICH (that also includes Intel MPI).

Add a test during configuration that checks whether the compination of
Fortran compiler and MPI implementation supports the usage of
`MPI_IN_PLACE`.
Define a preprocessor variable if it isn't supported.