Minor cleanup for CMake rules around VTK in ElmerGUI.
Do not hard-code paths that are build system dependent in the build
rules. They should be set by the user if necessary instead.
Remove some duplicate settings from ElmerGUI/CMakeLists.txt that are also
set in ElmerGUI/Application/CMakeLists.txt.
Don't include `VTK_USE_FILE` for versions of VTK in which it is deprecated.