Build ElmerIce modules in same directory as the ElmerFEM modules.
The libraries built by ElmerIce are used as modules for the ElmerSolver
library. Link them in the same directory as the other ElmerSolver
modules.

This allows to run more ctests of ElmerIce in the build tree on
Windows. It also makes the file layout during the ctests more similar to
the layout after installation.