ARPACK: Avoid compiler error with GFortran 10 or newer.
The `?vout` functions expect their third argument to be of rank-1. Newer
compilers (e.g., GFortran 10 or newer) emit an error if it is not.
That error is currently downgraded to a warning by adding
`-fallow-argument-mismatch` to the Fortran compiler flags for GFortran.

Avoid that error/warning by passing the respective argument with the
expected rank.