CI: Add timeout to ctests for the "quick" label.
Reduce the timeout for a single test from 1500 seconds (25 minutes is the
default for ctest) to 180 seconds.
None of the tests with the "quick" label should be taking longer than
that. And it isn't worth blocking the CI runners for any longer if it is
likely the case that the test would be stuck indefinitely.