Make building with UMFPACK optional.
Add new CMake variable `WITH_UMFPACK` which can be used to build without
using any functions from the UMFPACK library.

The default for this new variable is `TRUE` so there is no difference in
the build configuration to before unless a user explicitly configures
with, e.g., `-DWITH_UMFPACK=OFF`.

This probably fixes the issue described in #578.