Conditionally skip tests that depend on the UMFPACK solver.
Some tests are failing when built with `WITH_UMFPACK=OFF` and `WITH_MUMPS=OFF`
with an error like the following in the logs:
```
ERROR:: GetMatrixFormat: UMFPACK (and MUMPS) solver has not been installed.
```

Skip all tests conditional on `HAVE_UMFPACK OR HAVE_MUMPS` for which the string
"UMFPACK (and MUMPS) solver has not been installed." was found in the logs.