Avoid running tests using the same directory concurrently.
Some tests are failing intermittently in CI. They are passing when they
are re-run.
It looks like this is happening for tests which have a list of NPROCS
with more than one element. In that case, it might happen that multiple
instances of ElmerGrid are started in the same directory (if `ctest` is
running with parallel jobs). This might lead to the (intermittent) issue
that an ElmerFEM process is trying to read a grid folder that is
currently being written to by an ElmerGrid process from a test running
in parallel.

Avoid that potential concurrency issue by adding test dependencies
between the tests that are running in the same directory so that not more
than one test can run in the same directory at a time.