Set number of OpenMP threads to something reasonable in tests
Determine the number of physical processors of the build host and limit
the number of OpenMP threads to that for each test.
If a test is involving multiple MPI processes, divide the number of
physical processors by the number of used MPI processes. Set the number
of OpenMP threads to the integer floor of that result.

If a user sets `OMP_NUM_THREADS` explicitly in the environment, that
takes precedence over the automatic selection.